requirements.txt
dj-database-url==0.5.0
Django==1.11.14
django-ckeditor==5.6.1
django-js-asset==1.1.0
Pillow==5.2.0
python-decouple==3.1
pytz==2018.5
whitenoise==3.3.1
